requirement.txt
omegaconf
pytorch_lightning
kornia
open_clip_torch